+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Renderer/ShadowRenderPass.cpp
@@ -212,11 +212,11 @@ void FShadowRenderPass::RenderPrimitive(FStaticMeshRenderData* RenderData, const
 
         BufferManager->UpdateConstantBuffer(TEXT("FSubMeshConstants"), SubMeshData);
 
-        if (!OverrideMaterials.IsEmpty() && OverrideMaterials.Num() >= MaterialIndex && OverrideMaterials[MaterialIndex] != nullptr)
+        if (!OverrideMaterials.IsEmpty() && OverrideMaterials.Num() >= static_cast<int32>(MaterialIndex) && OverrideMaterials[MaterialIndex] != nullptr)
         {
             MaterialUtils::UpdateMaterial(BufferManager, Graphics, OverrideMaterials[MaterialIndex]->GetMaterialInfo());
         }
-        else if (!Materials.IsEmpty() && Materials.Num() >= MaterialIndex && Materials[MaterialIndex] != nullptr)
+        else if (!Materials.IsEmpty() && Materials.Num() >= static_cast<int32>(MaterialIndex) && Materials[MaterialIndex] != nullptr)
         {
             MaterialUtils::UpdateMaterial(BufferManager, Graphics, Materials[MaterialIndex]->Material->GetMaterialInfo());
         }
